+++ CHANGELOG.md
@@ -1,17 +1,36 @@
-* Support dynamic `page_cache_directory` using a Proc, Symbol or callable
+## 1.1.0 (January 23, 2017)
 
-* Support instance level setting of `page_cache_directory`
+*   Support dynamic `page_cache_directory` using a Proc, Symbol or callable
 
-* Add support for Rails 5.0 and master
+    *Andrew White*
 
-# 1.0.2
+*   Support instance level setting of `page_cache_directory`
 
-* Fix load order problem with other gems.
+    *Andrew White*
 
-# 1.0.1
+*   Add support for Rails 5.0 and master
 
-* Add Railtie to set `page_cache_directory` by default to `public` folder. (Fixes #5)
+    *Andrew White*
 
-# 1.0.0
 
-* First Release
+## 1.0.2 (November 15, 2013)
+
+*   Fix load order problem with other gems.
+
+    *Rafael Mendonça França*
+
+
+## 1.0.1 (October 24, 2013)
+
+*   Add Railtie to set `page_cache_directory` by default to `public` folder.
+
+    Fixes #5.
+
+    *Žiga Vidic*
+
+
+## 1.0.0 (February 27, 2013)
+
+*   Extract Action Pack - Action Caching from Rails core.
+
+    *Francesco Rodriguez*, *Rafael Mendonça França*, *Michiel Sikkes*